Summary: Debugging and Tidying up code Comments: Listed below:
Debugging:
---------------
DefaultEnergyPredictor :- 
* Fixing values for CPU usage and memory usage, (ensuring they follow the correct value ranges)
* Ensuring the correct conversion between power and energy.
* making sure the energy values for vms are set correctly.
VM:- 
* setting the diskGb from long to double in set method.

Tidying Up:
---------------
* DefaultEnergyPredictorTest - Making the test more comprehensive and realistic.
* ZabbixDatasourceAdaptoerTest - Adding more output
* DataGatherer - Making a few fields final, suppressing sleep in loop warning
* LoadBasedDivision - organising imports.
* ZabbixDataSourceAdaptor - Ensuring the units of memory are cleanly converted.

+++ iaas/energy-modeller/src/main/java/eu/ascetic/asceticarchitecture/iaas/energymodeller/types/energyuser/VM.java
@@ -124,7 +124,7 @@ public class VM extends EnergyUsageSource {
      *
      * @param diskGb The disk space this vm has been allocated.
      */
-    public void setDiskGb(long diskGb) {
+    public void setDiskGb(double diskGb) {
         if (diskGb < 0) {
             throw new IllegalArgumentException("The amount of disk size must not be less than zero.");
         }
